src/discussions_embeddings/requirements.txt
python-dotenv
azure-cosmos
azure-identity
openai